+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - [#1726](https://github.com/lapce/lapce/pull/1726): Add more panel theme keys, apply hover first, then current item colour
 
 ### Features/Changes
+- [#1791](https://github.com/lapce/lapce/pull/1791): Add highlighting for scope lines
 - [#1767](https://github.com/lapce/lapce/pull/1767): Added CMake tree-sitter syntax highlighting
 - [#1759](https://github.com/lapce/lapce/pull/1759): Update C tree-sitter and highlight queries
 - [#1758](https://github.com/lapce/lapce/pull/1758): Replaced dlang syntax highlighting